CHG: interpolation_param_dict.
Name changed for clarity.
It is now passed to interpolate_traces, instead of breaking it up into its
kwargs, s.t. it can be saved with the results.
CHG: if post_response_interval is None, then it is set to halfway between the
response start and end.  Having a fixed time (0.5s) was causing a problem with
trials that didn't have a response time longer than 0.5s (i.e., the animal did not
wait for reward).