Major overhaul of add_TTL_trial_start_times_to_behav_data.
It uses the same approach as reconcile_TTL_and_behav_trial_start_times,
but there were problems with that implementation.  This solution seems
more solid in its implementation (e.g., the reconcile solution used
gap indices from the TTL, but then inserted the INDEX into the
TIMESTAMP array that only has the timestamp times, not the full
time axis, so this seems like a bug).

This implementation also checks that there were indeed gaps where
there are mismatches.

It also checks that the timestamps are now synced and raises an
exception if not.

Adds flags for missing TTLs as in reconcile function, which can
later be used to filter these trials out.