FIX: TTL and behav reconciliation.
In the case that an added ITI did not improve the matching, nothing
happened, which caused an infinite loop.  There is now a distinction
between a correction that improves the match, and ITIs that are just
added to the end, for instance.  This makes the assert at the end
of the block hard to predict, +/- 1.