FIX: gaps flags disturbed finding trial start sequences in TTL.
If the gap occured between pre_start and start_code, then the code
would not recognize the trial as valid.  Simply allow the gap
code to precede the start_code.  This should not be the case
that was originally thought of, which was a redundancy of the
start code, which could be distinguished by what preceded it.