Change the way behav_df is handled.
The behavior is not specific to a probe, so it makes no sense
to save a version for each probe.

Previously, trials without matching TTLs were removed from
behav_df before saving.  Instead, now behav_df is saved with the
TTL flags and an additional flag from the metadata, which says
whether the curator noticed an invalid period in the recording.
Trials are labeled True if they are valid and False if not.
The code that checks for missing TTLs, invalid trials, and
selects only choice trials was moved into a new function in the
behavioral_utils: select_choice_trials_w_TTLs().

For trace alignment, behav_df is loaded and immediately run
through this new 'select_choice_trials_w_TTLs' filter.

When loading the data objects, behav_df is loaded as before (but
from the preprocessing directory, instead of the probe directory,
so that only one copy is necessary) and is sent through the filter
only when the traces are loaded.  This is saved in self.choice_df.

Rename cbehav_df->choice_df for clarity.