ADD: valid_periods to metadata. UPD: paths.
Paths were changed to use a copy in one's personal directory on the server
instead of directly writing to the original data.  That way the
data stays intact.

Also, the save directory is now distinguished from the place where
the metadata lives.  Save directory is used, at the moment, only
for clustering output, so I changed it to clustering_results/.

Metadata now lives in the data directory instead of the share
directory.